+++ src/app/auth/email-verification/page.tsx
@@ -1,0 +1,11 @@
+import EmailComponents from './components';
+
+const EmailPage = () => {
+  return (
+    <div className="w-[25rem] flex flex-col items-center pt-[3.75rem]">
+      <EmailComponents />
+    </div>
+  );
+};
+
+export default EmailPage;
