+++ src/app/auth/enter-email/page.tsx
@@ -1,11 +1,14 @@
-import EmailComponents from './components';
+import SignUpComponents from './components';
 
-const EmailPage = () => {
+const SignUpPage = () => {
   return (
     <div className="w-[25rem] flex flex-col items-center pt-[3.75rem]">
-      <EmailComponents />
+      <p className="font-body-14 mb-12 text-text-strong">
+        사용 중이신 이메일 주소로 가입할 수 있어요.
+      </p>
+      <SignUpComponents />
     </div>
   );
 };
 
-export default EmailPage;
+export default SignUpPage;
